Move state machine logic to a hook
