+++ src/components/App/App.js
@@ -1,37 +1,19 @@
 import React, { useEffect, useState } from "react";
-import { Machine } from "xstate";
-import { useMachine } from "@xstate/react";
 
 import Timer from "../Timer";
 import Controls from "../Controls";
+import useTimerControls from "./useTimerControls";
 
 const ONE_SECOND = 1000; // milliseconds
 const ONE_MINUTE = 60 * ONE_SECOND;
 const DEFAULT_LENGTH = ONE_MINUTE * 25;
 
-const stateTimerMachine = Machine({
-  id: "timer",
-  initial: "idle",
-  states: {
-    idle: {
-      on: {
-        START: "started"
-      }
-    },
-    started: {
-      on: { PAUSE: "paused", RESET: "idle" }
-    },
-    paused: {
-      on: { START: "started", RESET: "idle" }
-    }
-  }
-});
 function App() {
   const [currentTime, setCurrentTime] = useState(DEFAULT_LENGTH);
-  const [state, send] = useMachine(stateTimerMachine);
+  const { status, send } = useTimerControls();
 
   useEffect(() => {
-    if (state.value === "started") {
+    if (status === "started") {
       const updateTimer = setTimeout(() => {
         if (currentTime !== 0) {
           setCurrentTime(currentTime - ONE_SECOND);
@@ -40,10 +22,10 @@ function App() {
       return () => clearTimeout(updateTimer);
     }
 
-    if (state.value === "idle") {
+    if (status === "idle") {
       setCurrentTime(DEFAULT_LENGTH);
     }
-  }, [currentTime, state]);
+  }, [currentTime, status]);
 
   return (
     <main className="App">
@@ -53,7 +35,7 @@ function App() {
         onPlay={() => send("START")}
         onPause={() => send("PAUSE")}
         onReset={() => send("RESET")}
-        status={state.value}
+        status={status}
       />
     </main>
   );
